+++ src/EventStore.Core/Services/SubscriptionsService.cs
@@ -298,7 +298,7 @@ namespace EventStore.Core.Services
                 try
                 {
                     string[] parts = Helper.UTF8NoBom.GetString(eventRecord.Data).Split(_linkToSeparator, 2);
-                    int eventNumber = int.Parse(parts[0]);
+                    long eventNumber = long.Parse(parts[0]);
                     string streamId = parts[1];
 
                     var res = _readIndex.ReadEvent(streamId, eventNumber);
